requirements.txt
streamlit==1.41.1
httpx==0.27.2
python-dotenv==1.0.1
sqlglot==25.24.3
pydantic==2.10.4
pytest==8.3.4
